requirement.txt
flask
python-decouple
bs4
requests
mysql-connector-python